fix(emulator): u64 disk addressing — large (>4 GiB) devcontainer disks on wasm32 (CC-45)
The κ-disk capacity and byte addressing were `usize`. On wasm32 (the deployed browser
peer) `usize` is 32-bit, so a build-capable disk > 4 GiB wrapped: the guest-visible
virtio-blk capacity (`KappaBacking::len` → `capacity_sectors`) reported the size mod
2³² (an 8.25 GiB disk announced as 256 MiB → "bad geometry: block count … exceeds size
of device" → root-fs mount panic), and the request byte offset (`devbus`:
`(sector*512) as usize`) truncated any access past 4 GiB. Native (`usize` = 64-bit)
never showed it; the deployed amd64 path is where it bites.

Make the disk byte capacity and addressing `u64` throughout: `len`, `read_sector`,
`read_into`, `write_from`, the devbus blk request offset/bounds, and `capacity_sectors`
(no longer `as u64` after an already-truncated `usize`). The disk is now genuinely
parametric in size on every target — no hidden 4 GiB ceiling. Emulator unit tests
(incl. the 1 TiB occupancy index) stay green.

+++ crates/holospaces/src/emulator/devbus.rs
@@ -245,34 +245,36 @@ fn blk_service_chain(mem: &mut GuestRam, dev: &mut VirtioBlk, head: u16) -> u32 
     const VIRTIO_BLK_S_IOERR: u8 = 1;
 
     let mut written = 0u32;
-    let mut disk_off = (sector * 512) as usize;
+    // u64: the disk byte offset spans the full declared disk, which is multi-GiB and
+    // overflows a 32-bit usize on wasm32 (the deployed browser peer).
+    let mut disk_off = sector * 512;
     let mut status = VIRTIO_BLK_S_OK;
     match req_type {
         VIRTIO_BLK_T_IN => {
             for (addr, len, _flags) in data {
                 let n = *len as usize;
-                if disk_off + n > dev.disk.len() {
+                if disk_off + n as u64 > dev.disk.len() {
                     status = VIRTIO_BLK_S_IOERR;
                     break;
                 }
                 let mut buf = vec![0u8; n];
                 dev.disk.read_into(disk_off, &mut buf);
                 mem.write_bytes(*addr, &buf);
-                disk_off += n;
+                disk_off += n as u64;
                 written += *len;
             }
         }
         VIRTIO_BLK_T_OUT => {
             for (addr, len, _flags) in data {
                 let n = *len as usize;
-                if disk_off + n > dev.disk.len() {
+                if disk_off + n as u64 > dev.disk.len() {
                     status = VIRTIO_BLK_S_IOERR;
                     break;
                 }
                 let mut buf = vec![0u8; n];
                 mem.read_bytes(*addr, &mut buf);
                 dev.disk.write_from(disk_off, &buf);
-                disk_off += n;
+                disk_off += n as u64;
             }
         }
         VIRTIO_BLK_T_GET_ID => {
